2014_deptms: config file for enettv with a double cross validation associated to 'maskdep'

+++ 2014_deptms/maskdep/03_enettv_config_dCV.py
@@ -29,7 +29,7 @@ from scipy.stats import binom_test
 from collections import OrderedDict
 
 
-NFOLDS = 5
+NFOLDS = 10
 
 
 def load_globals(config):
@@ -38,8 +38,8 @@ def load_globals(config):
     GLOBAL.PARAMS = config["params"]
     GLOBAL.DILATATION = config['dilatation']
     GLOBAL.MAP_OUTPUT = config['map_output']
-    GLOBAL.OUTPUT_VALIDATION = config['output_validation']
     GLOBAL.OUTPUT_SELECTION = config['output_selection']
+    GLOBAL.OUTPUT_SUMMARY = config['output_summary']
     GLOBAL.PROB_CLASS1 = config["prob_class1"]
     GLOBAL.PENALTY_START = config["penalty_start"]
     STRUCTURE = nibabel.load(config["structure"])
@@ -95,7 +95,6 @@ def mapper(key, output_collector):
     ytr = GLOBAL.DATA_RESAMPLED_SELECTMODEL["y"][0]
     yval = GLOBAL.DATA_RESAMPLED_SELECTMODEL["y"][1]
     print key, "Data shape:", Xcalib.shape, Xte.shape, Xtr.shape, Xval.shape,
-    print ycalib.shape, yte.shape, ytr.shape, yval.shape
     STRUCTURE = GLOBAL.STRUCTURE
     #alpha, ratio_l1, ratio_l2, ratio_tv, k = key
     #key = np.array(key)
@@ -137,15 +136,16 @@ def reducer(key, values):
     #        for p in glob.glob("/neurospin/brainomics/2014_deptms/MRI/results/*/0.05_0.45_0.45_0.1_-1.0/")]
     # Compute sd; ie.: compute results on each folds
     criteria = {'recall_mean': [np.argmax, np.max],
-                'min_recall': [np.argmax, np.max], 
-                'max_pvalue_recall': [np.argmin, np.min]}
+                'min_recall': [np.argmax, np.max],
+                'max_pvalue_recall': [np.argmin, np.min],
+                'accuracy': [np.argmax, np.max],
+                'pvalue_accuracy': [np.argmin, np.min]}
     dilatation = GLOBAL.DILATATION
-    output_validation = GLOBAL.OUTPUT_VALIDATION
     output_selection = GLOBAL.OUTPUT_SELECTION
+    output_summary = GLOBAL.OUTPUT_SUMMARY
     map_output = GLOBAL.MAP_OUTPUT
     BASE = os.path.join("/neurospin/brainomics/2014_deptms/maskdep",
                         "results_enettv", dilatation, map_output)
-    print BASE
     INPUT = BASE + "/%i/%s"
     penalty_start = GLOBAL.PENALTY_START
     prob_class1 = GLOBAL.PROB_CLASS1
@@ -153,24 +153,22 @@ def reducer(key, values):
     keys = ['_'.join(str(e) for e in a) for a in params]
 
     compt = 0
-    if not os.path.isfile(output_validation):
+    if not os.path.isfile(output_selection):
         print "Model Construction, first cross-validation"
         # loop for the validation of the model
-        for fold in xrange(1, NFOLDS + 1):
-            print "fold: ", fold
+        for fold in xrange(0, NFOLDS + 1):
+            print "outer fold: ", fold
             # folds for the selection of the model associated to the test fold
-            idx_block = range((fold - 1) * (NFOLDS + 1) + 1,
-                              fold * (NFOLDS + 1))
-            print idx_block
+            idx_block = range(fold * (NFOLDS + 1),
+                              (fold + 1) * (NFOLDS + 1) -1)
             for key in keys:
                 print "key", key
-                print "compt: ", compt
                 paths_dCV = [INPUT % (idx, key) for idx in idx_block]
                 scores_CV = OrderedDict()
                 values = [GLOBAL.OutputCollector(p) for p in paths_dCV]
                 values = [item.load() for item in values]
                 n_fold = [item["n_fold"] for item in values]
-                assert n_fold == ([fold, fold, fold, fold, fold])
+                assert n_fold == ([fold for i in xrange(NFOLDS)])
                 recall_mean_std = np.std([np.mean(
                     precision_recall_fscore_support(
                     item["y_true"].ravel(), item["y_pred"])[1]) \
@@ -193,15 +191,25 @@ def reducer(key, values):
                 beta_cor_mean = np.mean(R[np.triu_indices_from(R, 1)])
                 success = r * s
                 success = success.astype('int')
+                accuracy = (r[0] * s[0] + r[1] * s[1])
+                accuracy = accuracy.astype('int')
                 pvalue_class0 = binom_test(success[0], s[0], 1 - prob_class1)
                 pvalue_class1 = binom_test(success[1], s[1], prob_class1)
+                pvalue_accuracy = binom_test(accuracy, s[0] + s[1], p=0.5)
                 k = key.split('_')
-                a, l1, tv = float(k[0]), float(k[1]), float(k[3])
+                a, l1 = float(k[0]), float(k[1])
+                l2, tv = float(k[2]), float(k[3])
+                k_ratio = float(k[4])
+                left = float(1 - tv)
+                if left == 0:
+                    left = 1.
                 scores_CV['n_fold'] = n_fold[0]
                 scores_CV['parameters'] = key
                 scores_CV['a'] = a
                 scores_CV['l1'] = l1
+                scores_CV['l2'] = l2
                 scores_CV['tv'] = tv
+                scores_CV['k_ratio'] = k_ratio
                 scores_CV['recall_0'] = r[0]
                 scores_CV['pvalue_recall_0'] = pvalue_class0
                 scores_CV['recall_1'] = r[1]
@@ -211,6 +219,8 @@ def reducer(key, values):
                                                             pvalue_class1)
                 scores_CV['recall_mean'] = r.mean()
                 scores_CV['recall_mean_std'] = recall_mean_std
+                scores_CV['accuracy'] = accuracy / float(s[0] + s[1])
+                scores_CV['pvalue_accuracy'] = pvalue_accuracy
                 scores_CV['precision_0'] = p[0]
                 scores_CV['precision_1'] = p[1]
                 scores_CV['precision_mean'] = p.mean()
@@ -230,11 +240,11 @@ def reducer(key, values):
                     scores_tab = pd.DataFrame(columns=scores_CV.keys())
                 scores_tab.loc[compt, ] = scores_CV.values()
                 compt += 1
-        scores_tab.to_csv(output_validation, index=False)
+        scores_tab.to_csv(output_selection, index=False)
 
-    if not os.path.isfile(output_selection):
+    if not os.path.isfile(output_summary):
         print "Model Selection"
-        scores_tab = pd.read_csv(output_validation)
+        scores_tab = pd.read_csv(output_selection)
         fold_groups = scores_tab.groupby('n_fold')
         compt = 0
         for fold_val, fold_group in fold_groups:
@@ -260,36 +270,8 @@ def reducer(key, values):
             scores_select_model.loc[compt, ] = scores_dCV.values()
             compt += 1
 
-        scores_select_model.to_csv(output_selection, index=False)
-    return None
-
-##############################################################################
-## Run all
-def run_all(config):
-    import mapreduce
-    BASE_PATH = "/neurospin/brainomics/2014_deptms"
-    OUTPUT_ENETTV = os.path.join(BASE_PATH,   "results_enettv")
-    params = config["params"][0]
-    key = '_'.join([str(p) for p in params])
-    modality = config["modality"]
-    roi = config["roi"]
-    WD = os.path.join(OUTPUT_ENETTV, modality + '_' + roi)
-    #class GLOBAL: DATA = dict()
-    load_globals(config)
-    OUTPUT = os.path.join(WD, 'test', key)
-    if not os.path.exists(OUTPUT):
-        os.makedirs(OUTPUT)
-    # run /home/ed203246/bin/mapreduce.py
-    oc = mapreduce.OutputCollector(OUTPUT)
-    X = np.load(os.path.join(WD,
-                             'X_' + modality + '_' + roi + '.npy'))
-    y = np.load(os.path.join(WD,  'y.npy'))
-    mapreduce.DATA_RESAMPLED = {}
-    mapreduce.DATA_RESAMPLED["X"] = [X, X]
-    mapreduce.DATA_RESAMPLED["y"] = [y, y]
-    params = np.array([float(p) for p in key.split("_")])
-    mapper(params, oc)
-    #oc.collect(key=key, value=ret)
+        scores_select_model.to_csv(output_summary, index=False)
+    return {}
 
 
 if __name__ == "__main__":
@@ -315,53 +297,77 @@ if __name__ == "__main__":
         elif dilat == "dilatation_within-brain_masks":
             dilat_process = "dilatation_within-brain"
         WD = os.path.join(ENETTV_PATH, dilat)
-
         if not os.path.exists(WD):
             os.makedirs(WD)
 
-        INPUT_DATA_X = os.path.join(DATASET_PATH, dilat,
-                                    'X_' + dilat_process + '.npy')
-        INPUT_DATA_y = os.path.join(DATASET_PATH, dilat,
-                                    'y.npy')
-        INPUT_MASK = os.path.join(DATASET_PATH, dilat,
-                                  'mask_' + dilat_process + '.nii')
         # copy X, y, mask file names in the current directory
-        shutil.copy2(INPUT_DATA_X, os.path.join(WD, 'X.npy'))
-        shutil.copy2(INPUT_DATA_y, WD)
-        shutil.copy2(INPUT_MASK, os.path.join(WD, 'mask.nii'))
+        if not os.path.isfile(os.path.join(WD, 'X.npy')):
+            INPUT_DATA_X = os.path.join(DATASET_PATH,
+                                    'X_' + dilat_process + '.npy')
+            shutil.copy2(INPUT_DATA_X, os.path.join(WD, 'X.npy'))
+        if not os.path.isfile(os.path.join(WD, 'y.npy')):
+            INPUT_DATA_y = os.path.join(DATASET_PATH, 'y.npy')
+            shutil.copy2(INPUT_DATA_y, os.path.join(WD, 'y.npy'))
+        if not os.path.isfile(os.path.join(WD, 'mask.nii')):
+            INPUT_MASK = os.path.join(DATASET_PATH,
+                                  'mask_' + dilat_process + '.nii')
+            shutil.copy2(INPUT_MASK, os.path.join(WD, 'mask.nii'))
 
         #################################################################
         ## Create config file
-        y = np.load(INPUT_DATA_y)
+        y = np.load(os.path.join(WD, 'y.npy'))
         prob_class1 = np.count_nonzero(y) / float(len(y))
 
         # resampling indexes for the double cross validation
         SEED_CV1 = 23071991
         SEED_CV2 = 5061931
         dcv = []
-        n_fold = 1
-        for calib, te in StratifiedKFold(y.ravel(), n_folds=NFOLDS,
-                                       shuffle=True, random_state=SEED_CV1):
-            assert((len(calib) + len(te)) == len(y))
-            y_calib = np.array([y[i] for i in calib.tolist()])
-            for val, tr in StratifiedKFold(y_calib.ravel(), n_folds=NFOLDS,
+        n_fold = 0
+        for i in xrange(2):
+            # if fold ext 0: y_calib and y_te = y
+            if n_fold == 0:
+                y_calib = y
+                y_te = y
+                calib = np.array([i for i in xrange(len(y))])
+                te = np.array([i for i in xrange(len(y))])
+                for val, tr in StratifiedKFold(y_calib.ravel(), n_folds=NFOLDS,
                                        shuffle=True, random_state=SEED_CV2):
-                assert((len(val) + len(tr)) == len(calib))
-                dcv.append([n_fold, calib.tolist(), te.tolist(),
+                    assert((len(val) + len(tr)) == len(calib))
+                    dcv.append([n_fold, calib.tolist(), te.tolist(),
                             val.tolist(), tr.tolist()])
-            dcv.append([n_fold, calib.tolist(), te.tolist(), None])
-            n_fold += 1
-        dcv.insert(0, None)  # first fold is None
+                # y_val and y_tr = y_calib
+                dcv.append([n_fold, calib.tolist(), te.tolist(), None])
+                n_fold += 1
+            else:
+                for calib, te in StratifiedKFold(y.ravel(), n_folds=NFOLDS,
+                                         shuffle=True, random_state=SEED_CV1):
+                    assert((len(calib) + len(te)) == len(y))
+                    y_calib = np.array([y[i] for i in calib.tolist()])
+                    for val, tr in StratifiedKFold(y_calib.ravel(),
+                                                   n_folds=NFOLDS,
+                                                   shuffle=True,
+                                                   random_state=SEED_CV2):
+                        assert((len(val) + len(tr)) == len(calib))
+                        dcv.append([n_fold, calib.tolist(), te.tolist(),
+                                    val.tolist(), tr.tolist()])
+                    # y_val and y_tr = y_calib
+                    dcv.append([n_fold, calib.tolist(), te.tolist(), None])
+                    n_fold += 1
 
         INPUT_DATA_X = 'X.npy'
         INPUT_DATA_y = 'y.npy'
         INPUT_MASK = 'mask.nii'
+
         # parameters grid
         # Re-run with
-        l1 = np.array([[0], [0.3], [0.7], [1]])
-        l1l2tv = np.hstack([l1, l1 - l1, 1 - l1])
-        alphas = [.01, .05, .1, .5, 1.]
+        tv = 0.3
+        ratios = np.array([[0.1, 0.9, 1], [0.9, 0.1, 1]])
+        alphas = [.05, .5]
         k_ratio = 1
+        l1l2tv = [np.array([[float(1 - tv),
+                             float(1 - tv),
+                             tv]]) * ratios]
+        l1l2tv = np.concatenate(l1l2tv)
         alphal1l2tv = np.concatenate([np.c_[np.array([[alpha]] * \
                                                        l1l2tv.shape[0]),
                                             l1l2tv] for alpha in alphas])
@@ -375,23 +381,26 @@ if __name__ == "__main__":
                       params=params, resample=dcv,
                       dilatation=dilat,
                       structure=INPUT_MASK,
-                      map_output="results_dCV",
+                      map_output="results_dCV_selection",
+                      map_output_validation="results_dCV_validation",
                       user_func=user_func_filename,
                       reduce_group_by="resample_index",
-                      output_validation="results_dCV.csv",
-                      output_selection="summary_selection.csv",
+                      output_selection="results_dCV_selection.csv",
+                      output_summary="summary_selection.csv",
+                      output_validation="results_dCV_validation.csv",
                       penalty_start=penalty_start,
                       prob_class1=prob_class1)
                       #reduce_output="results_dCV.csv")
-        json.dump(config, open(os.path.join(WD, "config_dCV.json"), "w"))
+        json.dump(config, open(os.path.join(WD, "config_dCV_selection.json"),
+                               "w"))
 
         #################################################################
-#        # Build utils files: sync (push/pull) and PBS
-#        import brainomics.cluster_gabriel as clust_utils
-#        sync_push_filename, sync_pull_filename, WD_CLUSTER = \
-#            clust_utils.gabriel_make_sync_data_files(WD)
-#        cmd = "mapreduce.py --map  %s/config_dCV.json" % WD_CLUSTER
-#        clust_utils.gabriel_make_qsub_job_files(WD, cmd)
+        # Build utils files: sync (push/pull) and PBS
+        import brainomics.cluster_gabriel as clust_utils
+        sync_push_filename, sync_pull_filename, WD_CLUSTER = \
+            clust_utils.gabriel_make_sync_data_files(WD)
+        cmd = "mapreduce.py --map  %s/config_dCV_selection.json" % WD_CLUSTER
+        clust_utils.gabriel_make_qsub_job_files(WD, cmd)
 #        ################################################################
 #        # Sync to cluster
 #        print "Sync data to gabriel.intra.cea.fr: "
